Add configdrive generation to Nova driver
This allows the nova driver to ask nova to create a configdrive for an
instance, compress it, base64 encoding it, and store it in the Ironic DB. Then the
agent will be able to download and decode the configdrive, then write it
to the disk. Configdrive allows Nova to set the root password, static
networking on an instance, and inject files.

Change-Id: I64f6ae7ff9403d99a9ff5d035a37d0ec5a3f74cf
(cherry picked from commit 75ec1b5ab75be78447876349cd933030abe78fff)