Add configdrive generation to Nova driver
This allows the nova driver to ask nova to create a configdrive for an
instance, base64 encoding it, and store it in the Ironic DB. Then the
agent will be able to download and decode the configdrive, then write it
to the disk. Configdrive allows Nova to set the root password, static
networking on an instance, and inject files.

Change-Id: I64f6ae7ff9403d99a9ff5d035a37d0ec5a3f74cf
(cherry picked from commit d1034e2a14ccc175c6e84152773ebbcf1be8f272)